+++ package.go
@@ -1,3 +1,12 @@
+/*
+Package owasp provides a checklist editor.
+
+This package allows developers to verify requirements with automated
+tests when possible. Also a simple markdown report generator is
+provided to quickly answer the question wether or not a specific level
+has been reached.
+
+*/
 package owasp
 
 // MustSetVerifiedNow loads and sets the verified flag. Panics on errors.